Core now supports named resources and resource 2nd stage convegence
Resources can now refer to each other by the heet base name preceeded
with an '@' character. As well, resources have a way to register a
converge_depedency() method that will run at program exit and can be
used to try to reconverge certain attributes that could not be converged
the first time around for some reason.