working past second converge
still has a bug in the normalizing of local rules.
we need to look up resource references in the normalization process
only SGs that can't be found should be normalized to resource
references.

Alternatively, when we get the remote rules, we need to normalize the
rules so that the source groups are normalized into resource references