Try to tag group twice on first creation
failing to tag the group that we just created (happens because of AWS
API inconsistencies) makes the group unable to be found on enext run of
the module, so the group then must be deleted manually.

for now, I just try to tag twice.
A better solution may be to try to look up by tag and if that fails, try
by name and then add the tag, or just ditch tags altogether, but I do
think an explicit tag is a good idea for later tools integrating w/ Heet