Fix CloudFormation updates
CAPABILITY_IAM was added to create, but never to update.
This fixes that in a brute-force way by just always adding the
capability to all API calls regardless of whether or not we need it.

TODO: add a method to iterate over the CF resources and determine if it
is actually needed rather than just this unconditional hotfix.